+++ CHANGELOG.md
@@ -78,6 +78,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `network:sta_status/0` to get the current connection state of the sta interface.
 - Added ESP32 `-DATOMVM_ELIXIR_SUPPORT=on` configuration option
 - Added support for ESP32 development builds to include NVS partition data at build time
+- Added support for new `is_integer/3` BIF, introduced with OTP-29
+- Support for OPT-29
 
 ### Changed